Move most of VersionUtility to VersionUtility2 in Tooling - Moved all of the compatibility checks - Moved all of the portable framework parsing code - Removed support for portable-x+y+z in project.json. Only profile fully spelled out portable profiles are supported.
