+++ src/Microsoft.Dnx.Tooling/NuGet/NETPortable/NetPortableProfileCollection.cs
