+++ src/Microsoft.Dnx.Tooling/NuGet/NETPortable/NetPortableProfileTable.cs
