+++ test/Microsoft.Dnx.Tooling.Tests/project.json
@@ -1,14 +1,15 @@
 {
-    "dependencies": {
-        "Microsoft.Dnx.CommonTestUtils": "1.0.0-*",
-        "Microsoft.Dnx.Tooling": "1.0.0-*",
-        "xunit.runner.aspnet": "2.0.0-aspnet-*"
-    },
-    "frameworks": {
-        "dnx451": { },
-        "dnxcore50": { }
-    },
-    "commands": {
-        "test": "xunit.runner.aspnet"
-    }
+  "dependencies": {
+    "Microsoft.AspNet.Testing": "1.0.0-*",
+    "Microsoft.Dnx.CommonTestUtils": "1.0.0-*",
+    "Microsoft.Dnx.Tooling": "1.0.0-*",
+    "xunit.runner.aspnet": "2.0.0-aspnet-*"
+  },
+  "frameworks": {
+    "dnx451": { },
+    "dnxcore50": { }
+  },
+  "commands": {
+    "test": "xunit.runner.aspnet"
+  }
 }
